requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Production Server
gunicorn==21.2.0

# Geocoding & Distance
geopy==2.4.1
haversine==2.8.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# API Documentation
drf-spectacular==0.27.0

# Environment Variables
python-dotenv==1.0.0
